Make SetShape not subtype ListShape
Subtyping SetShape from ListShape requires also subtyping
ListShape.Builder. However, to do that, the generic bounds of
SetShape.Builder in CollectionShape.Builder are now ListShape rather
than SetShape, and this can break previously written code doing fancy
stuff with type bounds (see the added test case).

This commit updates SetShape so it is not a subtype of ListShape, but it
is still considered a subtype in selectors, and it now offers isList ->
true, and asList returns a List. ShapeVisitor#setShape will call .asList
by default, though this can be confusing if the visitor uses things like
Shape.shapeToBuilder since it changes the shape type, and returning a
built ListShape at runtime from the set visitor will result in a class
cast exception.