Modify invalid range traits when needed
When loading 1.0 models, rather than dropping the default trait
from a member when the range trait of a shape is invalid for
its zero value, we now instead modify or drop the range trait and
keep the default trait. This ensures that the type doesn't change
in generated code.

Bump version to 1.25