Add default trait to root shapes and support nulls
This commit allows the default trait to be added to root level shapes which
then requires that any member that refers to the shape repeats the default
value _or_ overrides the default by setting the default trait to null, causing
the member to be optional. This provides full compatibility with Smithy IDL
v1 models where things like PrimitiveInteger's default could be overridden
by a box trait on a member.

PrimitiveInteger and related shapes in the prelude are no longer deprecated
and now have a `@default` trait applied to them set to the zero value of the
type, giving full compatibility with the semantics of Smithy IDL v1, but
forcing the default to be repeated on members removes the spooky action at
a distance previously observed with these primitive value bearing shapes.